Connected auth service to orchestratorManagementApp
Redid the orchestratorServiceManagement to have a express-based backend
that connects with the auth server and makes sure users are logged in
before using this app. It demonstrates the ability of our oAuth. Updated
each grunt file to actually work. Fixed grunt-hub to launch the whole
saas by just typing grunt.Made it so cpu usage is representative of node-
runner AND its child.